Refactor import settings
+ Move settings into their own class
+ Unify rake tasks; decide where to import from in Settings
