+++ app/importers/data_transformers/json_transformer.rb
@@ -1,9 +1,9 @@
 class JsonTransformer
 
   def transform(file)
-  	require 'json'
-  	parsed = JSON.parse(file, symbolize_names: true)
-  	parsed.each { |row| row[:local_id] = row[:local_id].to_s }
+    require 'json'
+    parsed = JSON.parse(file, symbolize_names: true)
+    parsed.each { |row| row[:local_id] = row[:local_id].to_s }
   end
-  
+
 end
