--- lib/tasks/import_kipp.rake
@@ -1,4 +1,0 @@
-task :import_kipp => :environment do
-  config = ImportInitializer.import_kippnj_config
-  ImportInitializer.import(config)
-end
